extracting only the copy function instead of the entire component
